move enum to common header file
